Actividad 1/server/requirements.txt
grpcio==1.24.3
grpcio-tools==1.24.3
regex==2018.1.10
requests==2.20.0